More scanner refactoring
